requirements.txt
# BTC Perpetual Elite Trader - Dependências
# Sistema de trading algorítmico avançado

# =============================================================================
# CORE DEPENDENCIES
# =============================================================================
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
joblib>=1.3.0

# =============================================================================
# MACHINE LEARNING & DEEP LEARNING
# =============================================================================
xgboost>=2.0.0
lightgbm>=4.1.0
catboost>=1.2.0
torch>=2.1.0
torchvision>=0.16.0
tensorflow>=2.13.0
keras>=2.13.0

# Ensemble e otimização
optuna>=3.4.0
hyperopt>=0.2.7
mlxtend>=0.23.0

# =============================================================================
# DATA PROCESSING & ANALYSIS
# =============================================================================
polars>=0.20.0
pyarrow>=14.0.0
duckdb>=0.9.0
h5py>=3.9.0
tables>=3.8.0

# Time series
statsmodels>=0.14.0
arch>=6.2.0
pmdarima>=2.0.0

# =============================================================================
# TECHNICAL ANALYSIS
# =============================================================================
pandas-ta>=0.3.14b
ta>=0.10.0
ta-lib>=0.4.0
finta>=1.3.0

# =============================================================================
# CRYPTOCURRENCY & TRADING
# =============================================================================
python-binance>=1.0.19
ccxt>=4.1.0
cryptofeed>=2.4.0
websocket-client>=1.6.0
websockets>=11.0.0

# =============================================================================
# PERFORMANCE & OPTIMIZATION
# =============================================================================
numba>=0.58.0
cython>=3.0.0
bottleneck>=1.3.0
numexpr>=2.8.0

# Parallel processing
multiprocessing-logging>=0.3.4
concurrent-futures>=3.1.1

# =============================================================================
# WEB FRAMEWORK & API
# =============================================================================
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
flask-sqlalchemy>=3.1.0
gunicorn>=21.2.0

# Dashboard e visualização
dash>=2.14.0
plotly>=5.17.0
bokeh>=3.3.0

# =============================================================================
# DATABASE & STORAGE
# =============================================================================
sqlalchemy>=2.0.0
alembic>=1.12.0
redis>=5.0.0
pymongo>=4.5.0

# =============================================================================
# NETWORKING & HTTP
# =============================================================================
aiohttp>=3.9.0
requests>=2.31.0
urllib3>=2.0.0
httpx>=0.25.0

# =============================================================================
# CONFIGURATION & UTILITIES
# =============================================================================
pyyaml>=6.0.1
python-dotenv>=1.0.0
configparser>=6.0.0
click>=8.1.0

# Date/time
python-dateutil>=2.8.0
pytz>=2023.3
pendulum>=2.1.0

# =============================================================================
# LOGGING & MONITORING
# =============================================================================
loguru>=0.7.0
structlog>=23.2.0
prometheus-client>=0.18.0

# =============================================================================
# TESTING & DEVELOPMENT
# =============================================================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
isort>=5.12.0

# =============================================================================
# VISUALIZATION & PLOTTING
# =============================================================================
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0
mplfinance>=0.12.0

# Interactive
ipywidgets>=8.1.0
jupyter>=1.0.0

# =============================================================================
# MATHEMATICAL & STATISTICAL
# =============================================================================
sympy>=1.12.0
cvxpy>=1.4.0
cvxopt>=1.3.0

# Risk management
pyfolio>=0.9.2
empyrical>=0.5.5
quantlib>=1.32.0

# =============================================================================
# ENCRYPTION & SECURITY
# =============================================================================
cryptography>=41.0.0
pycryptodome>=3.19.0
keyring>=24.2.0

# =============================================================================
# SYSTEM & OS
# =============================================================================
psutil>=5.9.0
schedule>=1.2.0
watchdog>=3.0.0

# =============================================================================
# OPTIONAL DEPENDENCIES (para funcionalidades específicas)
# =============================================================================

# GPU acceleration (opcional)
# cupy>=12.0.0  # Para NVIDIA GPUs
# cudf>=23.10.0  # GPU DataFrames

# Advanced ML (opcional)
# prophet>=1.1.4  # Facebook Prophet
# sktime>=0.24.0  # Time series ML
# tsfresh>=0.20.0  # Time series features

# Blockchain data (opcional)
# web3>=6.11.0  # Ethereum
# bitcoin>=1.1.42  # Bitcoin

# Alternative data (opcional)
# tweepy>=4.14.0  # Twitter API
# praw>=7.7.0  # Reddit API
# newspaper3k>=0.2.8  # News scraping

# =============================================================================
# DEVELOPMENT TOOLS
# =============================================================================
pre-commit>=3.5.0
bandit>=1.7.0
safety>=2.3.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# =============================================================================
# DEPLOYMENT
# =============================================================================
docker>=6.1.0
docker-compose>=1.29.0

# Cloud
boto3>=1.29.0  # AWS
google-cloud-storage>=2.10.0  # GCP
azure-storage-blob>=12.19.0  # Azure



# =============================================================================
# RETRY & RESILIENCE
# =============================================================================
tenacity>=8.2.0  # Para retry com back-off exponencial


# =============================================================================
# MONITORING & METRICS
# =============================================================================
prometheus_client>=0.16.0  # Para exportação de métricas Prometheus
psutil>=5.9.0  # Para métricas de sistema

# =============================================================================
# TESTING & QUALITY
# =============================================================================
pytest>=7.3.1
pytest-cov>=4.1.0
pytest-mock>=3.10.0
coverage>=7.2.7

